+++ Tests/New-PSDevelopmentProject.Tests.ps1
@@ -1,25 +1,25 @@
+﻿<#PSScriptInfo
 
-<#PSScriptInfo
-
-.VERSION 1.0.0
+.VERSION 1.1.0
 
 .GUID 00fe9040-1bc7-4862-9e45-e0a59c005d65
 
 .AUTHOR AdZero
 
-.COMPANYNAME 
+.COMPANYNAME AdZero
 
-.COPYRIGHT 
+.COPYRIGHT Copyright 2018-2021 AdZero
 
-.TAGS 
+.TAGS New-PSDevelopmentProject Test Pester
 
-.LICENSEURI 
+.LICENSEURI https://raw.githubusercontent.com/adzero/New-PSDevelopmentProject/master/LICENSE
 
-.PROJECTURI 
+.PROJECTURI https://github.com/adzero/New-PSDevelopmentProject
 
-.ICONURI 
+.ICONURI https://raw.githubusercontent.comm/adzero/New-PSDevelopmentProject/master/images/adzero-avatar.png
 
 .EXTERNALMODULEDEPENDENCIES 
+ Pester
 
 .REQUIREDSCRIPTS 
 
@@ -27,8 +27,9 @@
 
 .RELEASENOTES
 
-
 #>
+
+#Requires -Module Pester
 
 <# 
 
@@ -40,7 +41,7 @@ $here = (Get-Item (Split-Path -Parent $MyInvocation.MyCommand.Path)).Parent.Full
 $sut = (Split-Path -Leaf $MyInvocation.MyCommand.Path) -replace '\.Tests\.', '.'
 $name = (Get-Item (Split-Path -Parent $MyInvocation.MyCommand.Path)).Parent.Name
 
-$Script:scriptPath = "$here\$name\$sut"
+$Script:scriptPath = "$here\src\$sut"
 $Script:testPath = Join-Path -Path $env:TEMP -ChildPath ([Guid]::NewGuid().ToString())
 
 Describe "New-PSDevelopmentProject | Test script project creation" {
@@ -64,22 +65,22 @@ Describe "New-PSDevelopmentProject | Test script project creation" {
 		}
 		
 		It "Check test script project code directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src") -PathType Container| Should -Be $true
 		}
 
 		It "Check test script project code directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\TestScriptProject.ps1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\TestScriptProject.Format.ps1xml") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\lib") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\bin") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\TestScriptProject.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\TestScriptProject.Format.ps1xml") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\lib") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\bin") -PathType Container | Should -Be $true
 		}
 
 		It "Check test script project Pester tests directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\Tests") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\tests") -PathType Container| Should -Be $true
 		}
 
 		It "Check test script project Pester tests directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\Tests\TestScriptProject.Tests.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\tests\TestScriptProject.Tests.ps1") -PathType Leaf | Should -Be $true
 		}
 		
 		BeforeAll {
@@ -114,22 +115,22 @@ Describe "New-PSDevelopmentProject | Test script project creation" {
 		}
 		
 		It "Check test script project code directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src") -PathType Container| Should -Be $true
 		}
 
 		It "Check test script project code directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\TestScriptProject.ps1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\TestScriptProject.Format.ps1xml") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\lib") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\TestScriptProject\bin") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\TestScriptProject.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\TestScriptProject.Format.ps1xml") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\lib") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\src\bin") -PathType Container | Should -Be $true
 		}
 
 		It "Check test script project Pester tests directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\Tests") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\tests") -PathType Container| Should -Be $true
 		}
 
 		It "Check test script project Pester tests directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\Tests\TestScriptProject.Tests.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Scripts\TestScriptProject\tests\TestScriptProject.Tests.ps1") -PathType Leaf | Should -Be $true
 		}
 		
 		It "Check test script project git repository directory" {
@@ -172,27 +173,27 @@ Describe "New-PSDevelopmentProject | Test module project creation" {
 		}
 		
 		It "Check test module project code directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src") -PathType Container| Should -Be $true
 		}
 
 		It "Check test module project code directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.psm1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.psd1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.Format.ps1xml") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\Public") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\Private") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\lib") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\bin") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\en-US") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\en-US\about_TestModuleProject.help.txt") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.psm1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.psd1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.Format.ps1xml") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\Public") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\Private") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\lib") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\bin") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\en-US") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\en-US\about_TestModuleProject.help.txt") -PathType Leaf | Should -Be $true
 		}
 
 		It "Check test module project Pester tests directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\Tests") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\tests") -PathType Container| Should -Be $true
 		}
 
 		It "Check test module project Pester tests directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\Tests\TestModuleProject.Tests.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\tests\TestModuleProject.Tests.ps1") -PathType Leaf | Should -Be $true
 		}
 		
 		BeforeAll {
@@ -227,27 +228,27 @@ Describe "New-PSDevelopmentProject | Test module project creation" {
 		}
 		
 		It "Check test module project code directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src") -PathType Container| Should -Be $true
 		}
 
 		It "Check test module project code directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.psm1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.psd1") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\TestModuleProject.Format.ps1xml") -PathType Leaf | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\Public") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\Private") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\lib") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\bin") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\en-US") -PathType Container | Should -Be $true
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\TestModuleProject\en-US\about_TestModuleProject.help.txt") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.psm1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.psd1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\TestModuleProject.Format.ps1xml") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\Public") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\Private") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\lib") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\bin") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\en-US") -PathType Container | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\src\en-US\about_TestModuleProject.help.txt") -PathType Leaf | Should -Be $true
 		}
 
 		It "Check test module project Pester tests directory" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\Tests") -PathType Container| Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\tests") -PathType Container| Should -Be $true
 		}
 
 		It "Check test module project Pester tests directory default content" {
-			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\Tests\TestModuleProject.Tests.ps1") -PathType Leaf | Should -Be $true
+			Test-Path -Path (Join-Path -Path $testPath -ChildPath "Modules\TestModuleProject\tests\TestModuleProject.Tests.ps1") -PathType Leaf | Should -Be $true
 		}
 		
 		It "Check test module project git repository directory" {
@@ -272,49 +273,49 @@ Describe "New-PSDevelopmentProject | Test module project creation" {
 # SIG # Begin signature block
 # MIIIqQYJKoZIhvcNAQcCoIIImjCCCJYCAQExCzAJBgUrDgMCGgUAMGkGCisGAQQB
 # gjcCAQSgWzBZMDQGCisGAQQBgjcCAR4wJgIDAQAABBAfzDtgWUsITrck0sYpfvNR
-# AgEAAgEAAgEAAgEAAgEAMCEwCQYFKw4DAhoFAAQU3GXqEyqOvlzUJqolFeUbdQu2
-# zEqgggUwMIIFLDCCAxSgAwIBAgIQNC2Wu6czaZhGPNSKX+8pbjANBgkqhkiG9w0B
+# AgEAAgEAAgEAAgEAAgEAMCEwCQYFKw4DAhoFAAQUSyedCtgPu16a/grwg34ezcCM
+# rpigggUwMIIFLDCCAxSgAwIBAgIQW63XJ86VXrBOrf64HYKdVjANBgkqhkiG9w0B
 # AQ0FADAuMSwwKgYDVQQDDCNBZFplcm8gUG93ZXJTaGVsbCBMb2NhbCBDZXJ0aWZp
-# Y2F0ZTAeFw0xNzEwMTQyMjAwMDBaFw0xODEyMzAyMjAwMDBaMC4xLDAqBgNVBAMM
+# Y2F0ZTAeFw0yMDAxMDYyMzAwMDBaFw0yNTEyMzAyMzAwMDBaMC4xLDAqBgNVBAMM
 # I0FkWmVybyBQb3dlclNoZWxsIExvY2FsIENlcnRpZmljYXRlMIICIjANBgkqhkiG
-# 9w0BAQEFAAOCAg8AMIICCgKCAgEAseRnSUGLrhXeaxyhOFY40MKRLH7fS+HTPeEd
-# QijAvV7fMDRIVK/SovoUbHCz7pkf2lmgSStfVlTH44akTDwtzONVS2puwoLSYQBD
-# YcA5OAjEJTFwsUdMuZ8290vF3pnaeuOuiGvHbb5iMLSJ2OAxbiEAPHQONJdIhphF
-# vB00k+P7rnXdlQf6HbqcYWygUBF7qw/pTqn1o/u2uJdypw+NWUy/ybUP6i2dh/jb
-# 7uKHxhkHqmz2zHLfPYilQkHPj/uhswflgtrvwofrQTIP+43xwMueko3LNzKAM6Gu
-# 8Gh5Ojmv125nNCg8ghHJ5PcQ5taaYSodJ1UqZOs0TDPSoPg0ucXypV4zxTMFUVVZ
-# WoMG9ZRlVbbF1YQyqTonAiWpbe6FOYEYx78ozGW3BQp4QPzeDFsj+cXDRExnyeVJ
-# EsF10dTek94EeNvpdrWEFwgt7yTCKZRX791Kt9Yu1nnDgvb4CkyuKoXyYfC+Ne07
-# 3/QAg5gc1sMo62/50lBrZ4Mj1WABRza1zIlrKuzmEt9oE7XntGNn9MSfa57dc+mV
-# 7D5W5n/5i15S5sZTBu9/IjKrxkdeISUuIISJX/gF4dAdQp8G3OmCqRMsBfmau4xR
-# esqaTd+danE1qzHyMDteSc+eHtCGK6WKvKSKTB0K3bXso0Cp3zSUannkkWiXjwQZ
-# ZA+zQOECAwEAAaNGMEQwDgYDVR0PAQH/BAQDAgeAMBMGA1UdJQQMMAoGCCsGAQUF
-# BwMDMB0GA1UdDgQWBBT0Bf0PiCAkgYgmQ5hb7HM7Sx4nsjANBgkqhkiG9w0BAQ0F
-# AAOCAgEANQEi1mKrKsvhigzK/sIiy+12yVc4sS/GYU5Xyq9/qLYqYNYOihn7bJVW
-# cGHy+vdnoX8/cFdrTTyCFP2jUY2LckTnEwPbCWHnMEsuyVf3rftCV2zPw6tHpD3P
-# YFer5ipBdQxKs7/NOHYp8YEUghViKB0chDBhGvGZxqHDMARzPDQDzAR1hqp1tMzl
-# Hc6Mlk4IfvkmJE2qZh1K6Qg1x3NSJ4UnhmnkDWZmNJqbDsjV4WoNhaIQFoHaPbOT
-# 9YhESDW4T4WOogh4mxFDpueomroEimi31OlVRByYh5GkzKoaC3A6NXMdZ5R+YHeM
-# 7BQGZbzoFw4JP+6zA0o7BHlUV9a3whZt/r/a+zzu+Rq/1jVG8i8WhF0poSebB7/C
-# bxq1zJocAVJzLUY6QknhHIVgcM0wxz3g3afy5slGQxKKrvlQCBx0F6LdPsydxIl7
-# TH1W/iG1knjuH88CBvV9FyptvN15uzEzuhkbSUt6k5RVUfL7tsE0hprheFZJ5wwx
-# kSgvACunnmN8gudA0s0rbq65SsKpNMJjz8uAoxqyugoBbnbRV6sxt52qS10wX+5/
-# JS704Pqewi6qkescuWfW1QxrJOGzRD+UXNH+yplUlFlDQOcb7a5lwXrg+gQrSoQy
-# 0ZQiu6WZsYHX0L0BKEC+wT56x43RhmduEU63d6pzCz0wNcHOJaIxggLjMIIC3wIB
+# 9w0BAQEFAAOCAg8AMIICCgKCAgEAsKSGndXWnvczniCJM5x2ErFwKWPufBqG2hQT
+# NU/hxjQjOEfv7EewowFCf8hm0OQjNbn4Bv8LtwQCsN2r+iM+CScJ/mipyEpzvG9T
+# q6Hf4jybgBSYH8G5mWUym3LsrlFUt1A5FvfuJbPNNWkoGY6sgG7NTqEICLS46/Zc
+# n9GWNiYoIcMXdouMwWHsYLWnhKSfyE077brSmd4mJFym4OUy5tNiBjyiaEawZ6fE
+# vINXJghk2PfUUYjqBs/10AH75N8AjaBieBiQaZj98LAHJYis618Os/QxR4moRjGG
+# oMikBJMRWmC4ijrONFZsbchyxd/6gXLnUAcB1/F4g1VAXJZ3gPsRja1ItsxNuAcA
+# NEXRt34RKK+ayDul4mHvYy7X4+J88qB//p7ENrA+d1l2GM9GJhzemmgZUVv9Sx39
+# r45VF7zRRluCIRsmOREAKmiqRU6y6xdSp/Fmf0cWt2bBgvqJ5+j4cRHv4KI0veLP
+# SBWRFplbr+emxrhbMHI9m+/RHiT2BI9jBX9awXnaxcYH8IQN87CA/L4y/AhRLsuU
+# +zwaZ5YCyyi3VeAUfMCnJq/1Gfa+gUO2xfWcavVjcSgMatrktSNyCLURSD9OS/89
+# 4kBS2EO5lzvOUKDrE3LfDHORsbhpk7MbP/+TOnSL6heBplT6RE1Voql1k0IoTqyW
+# TtD1YA0CAwEAAaNGMEQwDgYDVR0PAQH/BAQDAgeAMBMGA1UdJQQMMAoGCCsGAQUF
+# BwMDMB0GA1UdDgQWBBRkZ706baqI0oKHJiJwnYJW+FT2rDANBgkqhkiG9w0BAQ0F
+# AAOCAgEAGkzJAIBdoCtUDV9Hz4bqllZ5X4SEEFk3WvlwqLz7S5FyiRFGzaBRfIoJ
+# XTwzczUXxr8UOhN90zBvkPoF+9pJccIEoXs3VuBK/jiv+WQJZtD2qpvmz5ZJMU3G
+# f/mwlDAc/Vof3GJmCIJ/5A4gVRu4j07tH/XWOyfYJEMIBPFOUsBnJCYZzlbKGXoU
+# hDQHLwsDB5Z/+4TwLZPSG3fkUewuTyaqLPotarh6EEWcf6Bxrtwr2SEIwpbshRb5
+# 1T8e2JRDAPFMM49kVqv5IiHq3Zrws4LFDZsXCuYaDABw4B7nDx1GP8En7+hGlVvn
+# Jhr1kr11yrwo4yr9RPvDLIQRNFrvkwwEcwBrGTXuydCNkd+P+knCDLR7T6B38i6o
+# WSiqleN0GgUYddT5s7kSPYjPbQD5ChheHYSTAiJBNvip+UBkTYsj9sxYz7sajmP0
+# vWhuXNqMgOj49KYJ3Q348Z93cMSBUZ4DYQoHToHtf9fXzHQGZtAOasDYayhUh08b
+# 1pr/zikZKfzH4Afgj5ffHLifxmWTUsocsIrXmkgScKDizW+vONhiljmS8FxacZmm
+# Xm+q+cDJ+2CrzlQSyPWN5f5r1DzyxC+7SA9uRfQi1meDSo0W9jrsKx+1IBE6D1DY
+# l/km38aF2oeAPBhn/43itNPYg2yP+nvWt/Fodn/t7BBk8upTPFYxggLjMIIC3wIB
 # ATBCMC4xLDAqBgNVBAMMI0FkWmVybyBQb3dlclNoZWxsIExvY2FsIENlcnRpZmlj
-# YXRlAhA0LZa7pzNpmEY81Ipf7yluMAkGBSsOAwIaBQCgeDAYBgorBgEEAYI3AgEM
+# YXRlAhBbrdcnzpVesE6t/rgdgp1WMAkGBSsOAwIaBQCgeDAYBgorBgEEAYI3AgEM
 # MQowCKACgAChAoAAMBkGCSqGSIb3DQEJAzEMBgorBgEEAYI3AgEEMBwGCisGAQQB
-# gjcCAQsxDjAMBgorBgEEAYI3AgEVMCMGCSqGSIb3DQEJBDEWBBQ4HCNnv7ymbozb
-# jDJrYdC5s+tm4DANBgkqhkiG9w0BAQEFAASCAgArLXpeYFHPOIBxp6aNoS8NRmiN
-# d2iMBK/3zNMEEQ7h8Onk6LdcjkAMBBF57IVSaEyp2LSEsAT32F/LoGP/HwVoIhUU
-# EcyHsWaDZ8ZfSaofC/4YBEEaXiG4o5p8Pnff1yWaRQD0DL1gdAC/U7pA9H55yNzJ
-# VXa3iZFsM8ulHI07gKsrCkNdorIFiBS75GuYggTRjc3ZoKnEX9cU61rHbqcwZobr
-# XEQrQ1oP9lXFYta5OwMWek8SlJdVxqD1tA1MFKZMuRX9bJprakioPjtutgPsEk2f
-# WLQdVKVSqJ0E9CEOZdAdvX19YlAirNO3018uzF+YRYbiwb1dl4bRGidQs+Df1lxI
-# Zst9a7cevByMa9U/d99brQuYfJ7kYIyRpa4FdzpXv3RqCcWsxXiWIe8xcTk9Sp2b
-# AKHX73nOgB7bvcx0blaG/uurJPz5OOT1HyANXTXQnWPstoRfcfx/jMV5JPkQfHWx
-# PekVcCamOBzCq8ZJWH/AJBCApEI7W2aQEjeJk4DvOQMonGnenJCj1+B7+Fm7+buI
-# LZYGx1raWxKx2dEyNT/6lTQh35vnMzHfKxsoJdaxW5hQYD07f2MtwgW0QksiReJ/
-# XnrCdUf7XYSwXJhV79AiJm2lOQTnNFhg447ZufXW+Df6gfr3a+5Afz05ZE8IVjW/
-# gVCEZCpwHeaNc2Otww==
+# gjcCAQsxDjAMBgorBgEEAYI3AgEVMCMGCSqGSIb3DQEJBDEWBBQ3/60AY3G46KjZ
+# vvJpes31yEkgODANBgkqhkiG9w0BAQEFAASCAgATwiP4f5ep9WgpzH+rlfCmgZrQ
+# AAXO4CFu8RUqr+SK/UXlSOf3djynr4TkozOZ51TaJFJguOC1WzrdNjCvmO9o1nmi
+# uBTmXOIimbCUkpXhZoK6/hffGSAcH3Pqobb54DAtzQHtsnw2OhkYWqrGvab9vcay
+# kLi+TIqFyH4Zsmf2Q3uVNQqC4ZFUBDRDfuvVTeAkvqDszeB94oWThTC3w89x7gJh
+# QkqXR7iKkKT/VGRSsiGtuSzE1wMEgmKq4fEZ1zAw50dOLSrIwj3vJ/G6hCtkotwm
+# Rl8kPrTzBPB2B3/CQa+iO4QP/+dlf/kThmQw9YCoVSztvK1Ag4itHosQKQfZUS0S
+# LT5ITNzSiEi5ZJSRlVKSF8nJ15PVqKeGBj6nlNUxAHRaMKSkLIJv5vrUlcHiLUkx
+# pneqz2XwL5fsAWf2G6zV5+uBX7ik+zb5E0Mt2e7IrL1aLodnowLxn/LDJ0NBlEhf
+# RiDsHmw6BxuhEERocbi1BhkCJ23sflIkIt3OkB3fWupXbMTkJIA13xshkPzlFhJP
+# 9bmLq/7KI9bDrAHUIjWlCbN8HsHNlB4alDSLTWd45+/aG6QsabbNUEhCCWJhXEJ3
+# initOSFHzAltWCXQkALCVQ5UbyaX5rF1gnpf8H/UoRvqwZEKKSEcCB4q7YVSrkkL
+# KN8QJLIv0DDaxHfNAA==
 # SIG # End signature block