+++ CHANGELOG.md
@@ -1,16 +1,30 @@
-# Changelog
+﻿# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2021-03-02
 
-## [1.0.1] - 2018-06-22
 ### Added
-- Changelog to track changes through versions
+
+- Option to disable automatic inclusion of the root folder corresponding to chosen project type.
+- Export of private functions for a module if ExportPrivate variable is et to true at module import.
+- Added script comment based help.
 
 ### Changed
+
+- Source and tests folders renamed to src and tests.
+- Git repository 'dev' branch name changed with 'develop'.
+
+## [1.0.1] - 2018-06-22
+
+### Added
+
+- Changelog to track changes through versions.
+
+### Changed
+
 - Default ps1xml format definition file contains XML declaration and types root node.
 
 ## [1.0.0] - 2018-06-16